requirements.txt
pygame
moderngl
PyOpenGL
PyOpenGL_accelerate